feat: token expiration (#107)
all tokens now have a lifetime limit

---------

+++ mlflow_oidc_auth/repository/user.py
@@ -1,10 +1,11 @@
-from typing import List, Optional, Callable
-from sqlalchemy.orm import Session
+from datetime import datetime, timezone
+from typing import Callable, List, Optional
 
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import RESOURCE_ALREADY_EXISTS, RESOURCE_DOES_NOT_EXIST
 from mlflow.utils.validation import _validate_username
 from sqlalchemy.exc import IntegrityError
+from sqlalchemy.orm import Session
 from werkzeug.security import check_password_hash, generate_password_hash
 
 from mlflow_oidc_auth.db.models import SqlUser
@@ -58,6 +59,7 @@ class UserRepository:
         self,
         username: str,
         password: Optional[str] = None,
+        password_expiration: Optional[datetime] = None,
         is_admin: Optional[bool] = False,
         is_service_account: Optional[bool] = False,
     ) -> User:
@@ -67,6 +69,8 @@ class UserRepository:
             user = get_user(session, username)
             if password is not None:
                 user.password_hash = generate_password_hash(password)
+            if password_expiration is not None:
+                user.password_expiration = password_expiration
             if is_admin is not None:
                 user.is_admin = is_admin
             if is_service_account is not None:
@@ -77,6 +81,8 @@ class UserRepository:
     def delete(self, username: str) -> None:
         with self._Session() as session:
             user = get_user(session, username)
+            if user is None:
+                raise MlflowException(f"User '{username}' not found.")
             session.delete(user)
             session.flush()
 
@@ -84,6 +90,11 @@ class UserRepository:
         with self._Session() as session:
             try:
                 user = get_user(session, username)
+                if user.password_expiration is not None:
+                    if user.password_expiration.tzinfo is None:
+                        user.password_expiration = user.password_expiration.replace(tzinfo=timezone.utc)
+                    if user.password_expiration < datetime.now(timezone.utc):
+                        return False
                 return check_password_hash(getattr(user, "password_hash"), password)
             except MlflowException:
                 return False
